requirements.txt
# Django Framework
Django==5.2.6

# Environment Variables
python-dotenv==1.0.0

# Web Server
gunicorn==21.2.0

# HTTP Requests
requests==2.31.0

# Additional packages (프로젝트에서 사용 중인 경우 추가)
# Pillow==10.1.0  # 이미지 처리
# psycopg2-binary==2.9.9  # PostgreSQL
# django-cors-headers==4.3.1  # CORS

psycopg[binary]>=3.1.8
pgvector
google-generativeai